Moved class TumorOnlyReportWorker to submodule cppNGSD. Moved class RtfDocument to cppNGS.
