+++ src/cppNGS/RtfDocument.cpp
